# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local Scientific Event Manager System",
        "title": "Scientific Event Manager API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/article/add-authors": {
            "post": {
                "description": "Adds one or more co-authors to an existing article",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Add co-authors to article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "articleID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "IDs of the co-authors, separated by commas",
                        "name": "coAuthorsID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ArticleResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Article/User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error updating co-authors",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/article/create": {
            "post": {
                "description": "While the user is logged in, creates an article and associates it to an event activity",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Create article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the associated event",
                        "name": "eventID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the associated activity",
                        "name": "activityID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of the article",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the article's first author",
                        "name": "firstAuthorID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The IDs of the co-authors, separated by commas",
                        "name": "coAuthorsID",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Publisher of the article",
                        "name": "publisher",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article's DOI",
                        "name": "doi",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Article's ISBN",
                        "name": "isbn",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "URL where the article is accessible",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.ArticleResponse"
                        }
                    },
                    "400": {
                        "description": "The activity does not belong to the event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User/event/activity not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error found during article creation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/article/delete": {
            "post": {
                "description": "While the user is logged in, deletes an article associated to an event which the user is the organizer",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Delete an article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the article",
                        "name": "articleID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "User is not the event organizer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Article/event/activity not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error found during article deletion",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/article/details": {
            "get": {
                "description": "Returns an article with its associated tags. Unpublished articles can only be viewed by the event organizer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Get article by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing article ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "You can't view this article (need to be the event organizer)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/article/list": {
            "get": {
                "description": "A user can view and filter the articles associated to an activity of a event that was published or which the user is the organizer",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "List the articles associated to an activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the associated activity",
                        "name": "activityID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter the name of the events shown",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "maximum": 50,
                        "type": "integer",
                        "default": 20,
                        "description": "Number of events shown",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of events to skip in the search",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Article"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Event was not published yet and the user is not the orgaziner",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No article found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Article": {
            "type": "object",
            "properties": {
                "coAuthorsID": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "doi": {
                    "type": "string"
                },
                "firstAuthorID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isbn": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "api.ArticleResponse": {
            "type": "object",
            "properties": {
                "articleID": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}